3.2 - Zmieniono lokalizację paczki command
