+++ src/main/java/edu/kis/powp/command/DriverCommand.java
@@ -1,0 +1,6 @@
+package edu.kis.powp.command;
+
+public interface DriverCommand {
+
+    void execute();
+}
