+++ src/main/java/edu/kis/powp/command/OperateToCommand.java
@@ -1,20 +1,19 @@
-package edu.kis.powp.jobs2d.drivers.command;
+package edu.kis.powp.command;
 
 import edu.kis.powp.jobs2d.Job2dDriver;
 
-public class SetPositionCommand implements DriverCommand{
+public class OperateToCommand implements DriverCommand {
 
     private Job2dDriver job2dDriver;
     private int startX = 0, startY = 0;
 
-    public SetPositionCommand(Job2dDriver job2dDriver, int startX, int startY) {
+    public OperateToCommand(Job2dDriver job2dDriver, int startX, int startY) {
         this.job2dDriver = job2dDriver;
         this.startX = startX;
         this.startY = startY;
     }
 
     @Override public void execute() {
-        job2dDriver.setPosition(startX,startY);
-        System.out.printf("SET");
+        job2dDriver.operateTo(startX,startY);
     }
 }
