--- src/main/java/edu/kis/powp/jobs2d/drivers/command/DriverCommand.java
@@ -1,8 +1,0 @@
-package edu.kis.powp.jobs2d.drivers.command;
-
-import edu.kis.powp.jobs2d.Job2dDriver;
-
-public interface DriverCommand {
-
-    void execute();
-}
